Add Lambda docker build command

+++ service/src/LambdaService.cpp
@@ -35,24 +35,44 @@ namespace AwsMock::Service {
     Dto::Lambda::CreateFunctionResponse LambdaService::CreateFunction(Dto::Lambda::CreateFunctionRequest &request) {
         poco_debug(_logger, "Create function request: " + request.ToString());
 
-        std::string codeDir;
-        if (!request.code.zipFile.empty()) {
-            codeDir = UnpackZipFile(request.code.zipFile);
+        // Build the docker image, if not existing
+        Dto::Docker::Image image;
+        if (!_dockerService->ImageExists(request.functionName, "latest")) {
+
+            // Unzip provided zip-file
+            std::string codeDir;
+            if (!request.code.zipFile.empty()) {
+                codeDir = UnpackZipFile(request.code.zipFile);
+            }
+
+            // Build the docker image
+            _dockerService->BuildImage(codeDir, request.functionName, "latest", request.handler);
+
+            // Get the image struct
+            image = _dockerService->GetImageByName(request.functionName +":latest");
+
+            // Cleanup
+            Core::DirUtils::DeleteDirectory(codeDir);
         }
 
-        // Build the docker image
-        std::string output = BuildDockerImage(codeDir, request.functionName, request.handler);
+        // Create the container, if not existing
+        std::string id = {};
+        if (!_dockerService->ContainerExists(request.functionName, "latest")) {
+            Dto::Docker::CreateContainerResponse containerCreateResponse = _dockerService->CreateContainer(request.functionName, "latest");
+            id = containerCreateResponse.id;
+        } else {
+            Dto::Docker::Container container = _dockerService->GetContainerByName(request.functionName, "latest");
+            id = container.id;
+        }
 
-        // Create ARN
-        std::string arn = Core::AwsUtils::CreateLambdaArn(_region, _accountId, request.functionName);
+        // Start container
+        _dockerService->StartContainer(id);
 
         // Create response
+        std::string arn = Core::AwsUtils::CreateLambdaArn(_region, _accountId, request.functionName);
         Dto::Lambda::CreateFunctionResponse
             response{.functionArn=arn, .functionName=request.functionName, .runtime=request.runtime, .role=request.role, .handler=request.handler,
             .environment=request.environment, .memorySize=request.memorySize};
-
-        // Cleanup
-        Core::DirUtils::DeleteDirectory(codeDir);
 
         return response;
     }
@@ -76,44 +96,6 @@ namespace AwsMock::Service {
         input.clear();
 
         return codeDir;
-    }
-
-    std::string LambdaService::BuildDockerImage(const std::string &codeDir, const std::string &name, const std::string &handler) {
-
-        std::string dockerFile = WriteDockerFile(codeDir, handler);
-
-        std::string imageFile = BuildImageFile(codeDir, name);
-
-        std::string tag = "latest";
-        std::string
-            header = Core::SystemUtils::SetHeader("POST", "/v1.42/build?t=" + name + ":" + tag + "&q=true", "application/x-tar", Core::FileUtils::FileSize(imageFile));
-        std::string output = Core::SystemUtils::SendFileViaDomainSocket("/var/run/docker.sock", header, imageFile);
-        poco_debug(_logger, "Docker image build, image: " + name + ":" + tag);
-        poco_debug(_logger, "Output: " + output);
-
-        return output;
-    }
-
-    std::string LambdaService::WriteDockerFile(const std::string &codeDir, const std::string &handler) {
-
-        std::string dockerFilename = codeDir + "Dockerfile";
-
-        std::ofstream ofs(dockerFilename);
-        ofs << "FROM public.ecr.aws/lambda/java:17" << std::endl;
-        ofs << "COPY classes/* ${LAMBDA_TASK_ROOT}/lib/" << std::endl;
-        ofs << "CMD [ \"" + handler + "::handleRequest\" ]" << std::endl;
-        poco_debug(_logger, "Dockerfile written, filename: " + dockerFilename);
-
-        return dockerFilename;
-    }
-
-    std::string LambdaService::BuildImageFile(const std::string &codeDir, const std::string &functionName) {
-
-        std::string tarFileName = codeDir + functionName + ".tgz";
-        Core::TarUtils::TarDirectory(tarFileName, codeDir);
-        poco_debug(_logger, "Gzip file written: " + tarFileName);
-
-        return tarFileName;
     }
 
 } // namespace AwsMock::Service
